+++ CHANGELOG.md
@@ -21,6 +21,7 @@
  - Removing methods `initialDimensions` and `removeGestureRecognizer` to avoid confusion
  - Adding standard for `SpriteComponent` and `SpriteAnimationComponent` constructors
  - Move files to comply with the dart package layout convention
+ - Fix gesture detection bug of children of `PositionComponent`
 
 ## 1.0.0-rc5
  - Option for overlays to be already visible on the GameWidget